stopping point

+++ controllers/sauce.js
@@ -25,6 +25,8 @@ exports.getSauce = (req, res, next) => {
 };
 
 exports.createSauce = (req, res, next) => {
+  if (!req.file) {
+  }
   const newSauce = req.body.sauce;
   const {
     name,
@@ -42,6 +44,7 @@ exports.createSauce = (req, res, next) => {
     name,
     manufacturer,
     description,
+    imageUrl: url + "/images/" + req.file.filename,
     mainPepper,
     heat,
     likes,
@@ -65,19 +68,7 @@ exports.createSauce = (req, res, next) => {
 
 exports.updateSauce = (req, res, next) => {
   const newSauce = req.body.sauce;
-  const {
-    name,
-    manufacturer,
-    description,
-    mainPepper,
-    heat,
-    likes,
-    dislikes,
-    userLike,
-    userDislikes,
-  } = newSauce;
-
-  const sauce = sauce
+  sauce
     .findById({ _id: req.params.id, newSauce }, newSauce)
     .then(() => {
       res.status(201).json({
@@ -91,11 +82,30 @@ exports.updateSauce = (req, res, next) => {
 
 exports.deleteSauce = (req, res, next) => {
   sauce
-    .deleteOne({ _id: req.params.id })
-    .then(() => {
-      res.status(201).json({
-        message: "Sauce deleted successfully",
-      });
+    .findOne({ _id: req.params.id })
+    .then((sauce) => {
+      if (!sauce) {
+        return res.status(404).json({
+          error: new Error("Sauce not found"),
+        });
+      }
+      if (sauce.userId !== req.auth.userId) {
+        return res.status(404).json({
+          error: new Error("Sauce not found"),
+        });
+      }
+      sauce
+        .deleteOne({ _id: req.params.id })
+        .then(() => {
+          res.status(201).json({
+            message: "Sauce deleted successfully",
+          });
+        })
+        .catch((error) => {
+          res.status(400).json({
+            error,
+          });
+        });
     })
     .catch((error) => {
       res.status(400).json({ error });
